gpt-oss-vllm-rag/pyproject.toml
[project]
name = "gpt-oss-vllm-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Davit Martirosyan",email = "davit@akna.tech"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph (==0.6.11)",
    "langchain-openai (==0.3.35)",
    "langchain-chroma (==0.2.6)",
    "langchain-community (==0.3.31)",
    "nltk (==3.9.1)",
    "vllm (==0.10.2)",
    "aiosqlite (==0.21.0)",
    "langgraph-checkpoint-sqlite (==2.0.11)",
    "langfuse (==3.9.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
notebook = "7.4.5"
ipywidgets = "8.1.7"

[[tool.poetry.source]]
name = "vllm-gpt-oss"
url = "https://wheels.vllm.ai/gpt-oss"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/nightly/cu128"
priority = "explicit"

[tool.poetry.dependencies]
torch = {source = "pytorch-gpu"}
torchvision = {source = "pytorch-gpu"}
torchaudio = {source = "pytorch-gpu"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
